Move common into a package
