+++ Go/Common/IFibonacciService.go
@@ -1,4 +1,4 @@
-package main
+package common
 
 type IFibonacciService interface {
 	GetNumbers(count int) []int
