+++ Go/LoopFivonacciService.go
@@ -1,5 +1,9 @@
 package main
 
+import (
+	"Fibonacci/common"
+)
+
 type LoopFibonacciService struct {
 }
 
@@ -8,8 +12,8 @@ func (s LoopFibonacciService) GetNumbers(count int) []int {
 
 	for i := 0; i < count; i++ {
 		if i > 0 {
-			var x, y int = getPriorValues(numbers)
-			var nextValue = getNextValue(x, y)
+			var x, y int = common.GetPriorValues(numbers)
+			var nextValue = common.GetNextValue(x, y)
 			numbers = append(numbers, nextValue)
 		} else {
 			numbers = append(numbers, 0)
